+++ examples/customize_connection.rs
@@ -1,13 +1,13 @@
 //! An example showing how to cutomize connections while using pooling.
 
-use async_bb8_diesel::{AsyncSimpleConnection, ConnectionError, DieselConnection};
+use async_bb8_diesel::{AsyncSimpleConnection, Connection, ConnectionError};
 use async_trait::async_trait;
 use diesel::pg::PgConnection;
 
 #[derive(Debug)]
 struct ConnectionCustomizer {}
 
-type DieselPgConn = DieselConnection<PgConnection>;
+type DieselPgConn = Connection<PgConnection>;
 
 #[async_trait]
 impl bb8::CustomizeConnection<DieselPgConn, ConnectionError> for ConnectionCustomizer {
@@ -20,7 +20,7 @@ impl bb8::CustomizeConnection<DieselPgConn, ConnectionError> for ConnectionCusto
 
 #[tokio::main]
 async fn main() {
-    let manager = async_bb8_diesel::DieselConnectionManager::<PgConnection>::new("localhost:1234");
+    let manager = async_bb8_diesel::ConnectionManager::<PgConnection>::new("localhost:1234");
     let _ = bb8::Pool::builder()
         .connection_customizer(Box::new(ConnectionCustomizer {}))
         .build(manager)
